malii-t6xx: Graphic Driver updated to rev 65537

+++ drivers/gpu/arm/t6xx/kbase/src/common/mali_kbase_js_policy.h
@@ -727,7 +727,6 @@ void kbasep_js_policy_deregister_job( kbasep_js_policy *js_policy, kbase_context
  *
  * The caller has the following conditions on locking:
  * - kbasep_js_device_data::runpool_lock::irq will be held.
- * - kbdev->jm_slots[ job_slot_idx ].lock will be held
  * - kbasep_js_device_data::runpool_mutex will be held.
  * - kbasep_js_kctx_info::ctx::jsctx_mutex. will be held
  */
@@ -759,7 +758,6 @@ mali_bool kbasep_js_policy_dequeue_job( kbase_device *kbdev,
  *
  * The caller has the following conditions on locking:
  * - kbasep_js_device_data::runpool_irq::lock will be held.
- * - kbdev->jm_slots[ job_slot_idx ].lock will be held
  *
  * @note The caller \em might be holding one of the
  * kbasep_js_kctx_info::ctx::jsctx_mutex locks, if this code is called from
@@ -820,7 +818,7 @@ void kbasep_js_policy_enqueue_job( kbasep_js_policy *js_policy, kbase_jd_atom *k
  * @param katom         job dispatch atom
  * @param time_spent_us the time spent by the job, in microseconds (10^-6 seconds).
  */
-void kbasep_js_policy_log_job_result( kbasep_js_policy *js_policy, kbase_jd_atom *katom, u32 time_spent_us );
+void kbasep_js_policy_log_job_result( kbasep_js_policy *js_policy, kbase_jd_atom *katom, u64 time_spent_us );
 
 /** @} */ /* end group kbase_js_policy_job */
